+++ Assets/Scripts/Enemy/StateMachine/State.cs
@@ -7,7 +7,7 @@ public class State : MonoBehaviour
 
     protected Player Target { get; private set; }
 
-    public virtual void StartBuff(float multiple, float actionTime) { }
+    public virtual void StartBuff(int multiple, float actionTime) { }
 
     public void Enter(Player target)
     {
